+++ fabric-common-tags-v1/src/datagen/java/net/fabricmc/fabric/impl/tag/datagen/DatagenEntrypoint.java
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.tag.common.datagen;
+package net.fabricmc.fabric.impl.tag.datagen;
 
-import net.fabricmc.fabric.impl.tag.common.datagen.generators.BiomeTagGenerator;
-import net.fabricmc.fabric.impl.tag.common.datagen.generators.BlockTagGenerator;
-import net.fabricmc.fabric.impl.tag.common.datagen.generators.EnchantmentTagGenerator;
-import net.fabricmc.fabric.impl.tag.common.datagen.generators.EntityTypeTagGenerator;
-import net.fabricmc.fabric.impl.tag.common.datagen.generators.FluidTagGenerator;
-import net.fabricmc.fabric.impl.tag.common.datagen.generators.ItemTagGenerator;
+import net.fabricmc.fabric.impl.tag.datagen.generators.BiomeTagGenerator;
+import net.fabricmc.fabric.impl.tag.datagen.generators.BlockTagGenerator;
+import net.fabricmc.fabric.impl.tag.datagen.generators.EnchantmentTagGenerator;
+import net.fabricmc.fabric.impl.tag.datagen.generators.EntityTypeTagGenerator;
+import net.fabricmc.fabric.impl.tag.datagen.generators.FluidTagGenerator;
+import net.fabricmc.fabric.impl.tag.datagen.generators.ItemTagGenerator;
 import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
