+++ fabric-common-tags-v1/src/datagen/java/net/fabricmc/fabric/impl/tag/datagen/generators/BlockTagGenerator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.tag.common.datagen.generators;
+package net.fabricmc.fabric.impl.tag.datagen.generators;
 
 import net.minecraft.block.Blocks;
 import net.minecraft.tag.BlockTags;
