+++ fabric-common-tags-v1/src/datagen/java/net/fabricmc/fabric/impl/tag/datagen/generators/FluidTagGenerator.java
@@ -14,33 +14,24 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.tag.common.datagen.generators;
+package net.fabricmc.fabric.impl.tag.datagen.generators;
 
-import net.minecraft.entity.EntityType;
+import net.minecraft.tag.FluidTags;
 
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
-import net.fabricmc.fabric.api.tag.v1.CommonEntityTypeTags;
+import net.fabricmc.fabric.api.tag.v1.CommonFluidTags;
 
-public class EntityTypeTagGenerator extends FabricTagProvider.EntityTypeTagProvider {
-	public EntityTypeTagGenerator(FabricDataGenerator dataGenerator) {
+public class FluidTagGenerator extends FabricTagProvider.FluidTagProvider {
+	public FluidTagGenerator(FabricDataGenerator dataGenerator) {
 		super(dataGenerator);
 	}
 
 	@Override
 	protected void generateTags() {
-		getOrCreateTagBuilder(CommonEntityTypeTags.BOSSES)
-				.add(EntityType.ENDER_DRAGON)
-				.add(EntityType.WITHER);
-		getOrCreateTagBuilder(CommonEntityTypeTags.MINECARTS)
-				.add(EntityType.MINECART)
-				.add(EntityType.TNT_MINECART)
-				.add(EntityType.CHEST_MINECART)
-				.add(EntityType.FURNACE_MINECART)
-				.add(EntityType.COMMAND_BLOCK_MINECART)
-				.add(EntityType.HOPPER_MINECART)
-				.add(EntityType.SPAWNER_MINECART);
-		getOrCreateTagBuilder(CommonEntityTypeTags.BOATS)
-				.add(EntityType.BOAT);
+		getOrCreateTagBuilder(CommonFluidTags.WATER)
+				.addOptionalTag(FluidTags.WATER);
+		getOrCreateTagBuilder(CommonFluidTags.LAVA)
+				.addOptionalTag(FluidTags.LAVA);
 	}
 }
